refactor: split richInputTile utilities out of contentEditable
Move tile-specific code (creation, adjacency detection, paste
thresholds, SVG icons) into its own module. contentEditable.ts
now only contains generic cursor, insertion, and text extraction
utilities.
